remove support for electron
