--- packages/electron/package.json
@@ -1,73 +1,0 @@
-{
-    "name": "electron-package",
-    "version": "0.0.1",
-    "dependencies": {
-        "browser-window": "^0.4.0",
-        "html-webpack-plugin": "3.2.0"
-    },
-    "devDependencies": {
-        "@babel/register": "7.8.3",
-        "devtron": "1.4.0",
-        "electron": "7.1.10",
-        "electron-builder": "22.2.0",
-        "electron-debug": "3.0.1",
-        "electron-devtools-installer": "2.2.4"
-    },
-    "browserslist": ["electron 1.6"],
-    "main": "../../build/electron/index.js",
-    "build": {
-        "appId": "SubstraFoundation.github.io",
-        "productName": "SubstraFoundation website",
-        "directories": {
-            "buildResources": "resources",
-            "output": "release"
-        },
-        "files": [
-            "../../build/electron/**/*"
-        ],
-        "linux": {
-            "target": [
-                "deb",
-                "AppImage"
-            ]
-        },
-        "mac": {
-            "target": "dmg"
-        },
-        "win": {
-            "target": "nsis"
-        },
-        "nsis": {
-            "artifactName": "${productName}_${version}_win.${ext}"
-        },
-        "dmg": {
-            "artifactName": "${productName}_${version}_mac.${ext}",
-            "contents": [
-                {
-                    "x": 130,
-                    "y": 220
-                },
-                {
-                    "x": 410,
-                    "y": 220,
-                    "type": "link",
-                    "path": "/Applications"
-                }
-            ]
-        }
-    },
-    "scripts": {
-        "dev": "cross-env IS_ELECTRON=1 START_HOT=1 npm run start-renderer-dev",
-        "start-main-dev": "cd ../../ && cross-env IS_ELECTRON=1 HOT=1 NODE_ENV=development electron -r @babel/register --require @babel/polyfill ./src/electron/main",
-        "start-renderer-dev": "cd ../../ && cross-env IS_ELECTRON=1 NODE_ENV=development babel-node ./node_modules/webpack-dev-server/bin/webpack-dev-server --config ./webpack/electron/renderer.dev.js",
-        "start:electron": "cd ../../ && cross-env IS_ELECTRON=1 NODE_ENV=production electron ./build/electron/index.js",
-        "build:electron": "concurrently \"npm run build-main\" \"npm run build-renderer\"",
-        "build-main": "cd ../../ && cross-env IS_ELECTRON=1 NODE_ENV=production babel-node ./node_modules/webpack/bin/webpack --progress --config webpack/electron/main.js --colors && cd -",
-        "build-renderer": "cd ../../ && cross-env IS_ELECTRON=1 NODE_ENV=production babel-node ./node_modules/webpack/bin/webpack --progress --config webpack/electron/renderer.prod.js --colors && cd -",
-        "package": "cd ../../ && npm run build:electron && build --publish never && cd -",
-        "package-all": "cd ../../ && npm run build:electron && build -mwl && cd -",
-        "package-linux": "cd ../../ && DEBUG_PROD=true npm run build:electron && build --linux && cd -",
-        "package-win": "cd ../../ && npm run build:electron && build --win --x64 && cd -",
-        "build-electron": "cd ../../ && build --publish never && cd -"
-    }
-}
