--- src/app/business/routes/notFound/routes/electron.js
@@ -1,4 +1,0 @@
-// import classic, no need to lazy load on desktop
-import Component from '../components/index';
-
-export default Component;
