--- src/app/business/routes/notFound/routes/index.js
@@ -1,1 +1,0 @@
-module.exports = process.env.IS_ELECTRON !== 'false' ? require('./electron.js') : require('./web.js');
