+++ src/common/configureStore/options.js
@@ -1,16 +1,11 @@
 import {redirect} from 'redux-first-router';
 import queryString from 'query-string';
-import {endsWith} from 'lodash';
 
 export default {
     onAfterChange: (dispatch, getState) => {
         const {user, location} = getState();
 
-        // for handling electron
-        if (endsWith(location.pathname, '/dist/index.html')) {
-            dispatch(redirect({type: 'HOME'}));
-        }
-        else if (location.type === 'LOGIN' && user && user.authenticated) {
+        if (location.type === 'LOGIN' && user && user.authenticated) {
             dispatch(redirect({type: 'HOME'}));
         }
         else if (location.type !== 'LOGIN' && user && !user.authenticated) {
