--- src/electron/app.ejs
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title><%= htmlWebpackPlugin.options.title %></title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <div id="devTools"></div>
-  </body>
-</html>
