--- src/electron/main.js
@@ -1,87 +1,0 @@
-import {app, BrowserWindow, globalShortcut} from 'electron';
-
-import MenuBuilder from './menu';
-
-let mainWindow = null;
-
-if (process.env.NODE_ENV === 'production') {
-    const sourceMapSupport = require('source-map-support');
-    sourceMapSupport.install();
-}
-
-if (process.env.NODE_ENV === 'development' || process.env.DEBUG_PROD === 'true') {
-    require('electron-debug')();
-    const path = require('path');
-    const p = path.join(__dirname, '..', '..', 'src', 'electron', 'node_modules');
-    require('module').globalPaths.push(p);
-}
-
-const installExtensions = async () => {
-    const installer = require('electron-devtools-installer');
-    const forceDownload = !!process.env.UPGRADE_EXTENSIONS;
-    const extensions = [
-        'REACT_DEVELOPER_TOOLS',
-        'REDUX_DEVTOOLS',
-    ];
-
-    return Promise
-        .all(extensions.map((name) => installer.default(installer[name], forceDownload)))
-        .catch(console.log);
-};
-
-
-/**
- * Add event listeners...
- */
-
-app.on('window-all-closed', () => {
-    // Respect the OSX convention of having the application in memory even
-    // after all windows have been closed
-    if (process.platform !== 'darwin') {
-        app.quit();
-    }
-});
-
-
-app.on('ready', async () => {
-    if (process.env.NODE_ENV === 'development' || process.env.DEBUG_PROD === 'true') {
-        await installExtensions();
-    }
-
-    mainWindow = new BrowserWindow({
-        show: false,
-        width: 1024,
-        height: 728,
-        webPreferences: {
-            nodeIntegration: true,
-        },
-    });
-
-    const port = process.env.PORT || 1212;
-
-    mainWindow.loadURL(process.env.NODE_ENV === 'development'
-        ? `http://electron.node-1.com:${port}/`
-        : `file://${__dirname}/dist/index.html`);
-
-    // @TODO: Use 'ready-to-show' event
-    //        https://github.com/electron/electron/blob/master/docs/api/browser-window.md#using-ready-to-show-event
-    mainWindow.webContents.on('did-finish-load', () => {
-        if (!mainWindow) {
-            throw new Error('"mainWindow" is not defined');
-        }
-        mainWindow.show();
-        mainWindow.focus();
-    });
-
-    globalShortcut.register('f5', () => {
-        console.log('f5 is pressed');
-        mainWindow.reload();
-    });
-
-    mainWindow.on('closed', () => {
-        mainWindow = null;
-    });
-
-    const menuBuilder = new MenuBuilder(mainWindow);
-    menuBuilder.buildMenu();
-});
