--- webpack/electron/main.js
@@ -1,96 +1,0 @@
-/**
- * Webpack config for production electron main process
- */
-
-import HappyPack from 'happypack';
-import merge from 'webpack-merge';
-import TerserPlugin from 'terser-webpack-plugin';
-import {BundleAnalyzerPlugin} from 'webpack-bundle-analyzer';
-import baseConfig from './base';
-
-import definePlugin from '../utils/plugins/definePlugin';
-import rules from '../utils/rules';
-
-export default merge.smart(baseConfig, {
-    devtool: 'source-map',
-    target: 'electron-main',
-    entry: './src/electron/main',
-    // 'main.js' in root
-    output: {
-        path: __dirname,
-        filename: '../../build/electron/index.js',
-    },
-    module: {
-        rules: rules(),
-    },
-    plugins: [
-        new HappyPack({
-            id: 'babel',
-            loaders: [{
-                path: 'babel-loader', // Options to configure babel with
-                query: {
-                    babelrc: false,
-                    plugins: [
-                        ['universal-import', {
-                            disableWarnings: true,
-                        }],
-                        'emotion',
-                        '@babel/plugin-transform-runtime',
-                        'lodash',
-                        'react-hot-loader/babel',
-
-                        // Stage 0
-                        '@babel/plugin-proposal-function-bind',
-
-                        // Stage 1
-                        '@babel/plugin-proposal-export-default-from',
-                        '@babel/plugin-proposal-logical-assignment-operators',
-                        ['@babel/plugin-proposal-optional-chaining', {loose: false}],
-                        ['@babel/plugin-proposal-pipeline-operator', {proposal: 'minimal'}],
-                        ['@babel/plugin-proposal-nullish-coalescing-operator', {loose: false}],
-                        '@babel/plugin-proposal-do-expressions',
-
-                        // Stage 2
-                        ['@babel/plugin-proposal-decorators', {legacy: true}],
-                        '@babel/plugin-proposal-function-sent',
-                        '@babel/plugin-proposal-export-namespace-from',
-                        '@babel/plugin-proposal-numeric-separator',
-                        '@babel/plugin-proposal-throw-expressions',
-
-                        // Stage 3
-                        '@babel/plugin-syntax-dynamic-import',
-                        '@babel/plugin-syntax-import-meta',
-                        ['@babel/plugin-proposal-class-properties', {loose: true}],
-                        '@babel/plugin-proposal-json-strings',
-                    ],
-                    presets: [
-                        ['@babel/preset-env', {modules: false}],
-                        '@babel/preset-react',
-                    ],
-                },
-            }],
-            threads: 4,
-        }),
-        new BundleAnalyzerPlugin({
-            analyzerMode: process.env.OPEN_ANALYZER === 'true' ? 'server' : 'disabled',
-            openAnalyzer: process.env.OPEN_ANALYZER === 'true',
-        }),
-        definePlugin(),
-    ],
-    optimization: {
-        minimizer: [new TerserPlugin({
-            cache: true,
-            parallel: true,
-        })],
-    },
-
-    /**
-     * Disables webpack processing of __dirname and __filename.
-     * If you run the bundle in node.js it falls back to these values of node.js.
-     * https://github.com/webpack/webpack/issues/2010
-     */
-    node: {
-        __dirname: false,
-        __filename: false,
-    },
-});
